+++ src/Terrabuild.Extensions.Tests/Scripts/Helpers.fs
@@ -84,4 +84,5 @@ let normalizeOps (ops: ShellOperations) =
 let op command arguments errorLevel =
     { ShellOperation.Command = command
       ShellOperation.Arguments = arguments
-      ShellOperation.ErrorLevel = errorLevel }
+      ShellOperation.ErrorLevel = errorLevel
+      ShellOperation.Stdout = None }