+++ src/Terrabuild.Scripting/Contracts.fs
@@ -53,6 +53,7 @@ module ScriptingContracts =
         Command: string
         Arguments: string
         ErrorLevel: int
+        Stdout: string option
     }
 
     type ShellOperations = ShellOperation list